+++ CHANGELOG.rst
@@ -2,6 +2,14 @@
 Version History
 ===============
 
+
+
+v6.6.0
+------
+
+* Remove authlist references `<https://github.com/lsst-ts/LOVE-producer/pull/148>`_
+* Refactor ScriptQueueState payload into several `<https://github.com/lsst-ts/LOVE-producer/pull/147>`_
+
 v6.5.4
 ------